+++ linuxdoom-1.10/r_main.c
@@ -95,7 +95,7 @@ angle_t			clipangle;
 // maps the visible view angles to screen X coordinates,
 // flattening the arc to a flat projection plane.
 // There will be many angles mapped to the same X. 
-int			viewangletox[FINEANGLES/2];
+short			viewangletox[FINEANGLES/2];
 
 // The xtoviewangleangle[] table maps a screen pixel
 // to the lowest viewangle that maps back to x ranges
@@ -544,8 +544,8 @@ void R_InitTables (void)
 void R_InitTextureMapping (void)
 {
     int			i;
-    int			x;
-    int			t;
+    short		x;
+    int 		t;
     fixed_t		focallength;
     
     // Use tangent table to generate viewangletox:
