+++ linuxdoom-1.10/r_state.h
@@ -81,7 +81,7 @@ extern vertex_t*	vertexes;
 extern int		numsegs;
 extern seg_t*		segs;
 
-extern int		numsectors;
+extern short		numsectors;
 extern sector_t*	sectors;
 
 extern int		numsubsectors;
@@ -111,7 +111,7 @@ extern player_t*	viewplayer;
 // ?
 extern angle_t		clipangle;
 
-extern int		viewangletox[FINEANGLES/2];
+extern short		viewangletox[FINEANGLES/2];
 extern angle_t		xtoviewangle[SCREENWIDTH+1];
 //extern fixed_t		finetangent[FINEANGLES/2];
